Properly handle optional chunkSize in AIMultipleChooserPanel
Previously, the data attribute will be set to 'None'.

Then, when the value is accessed in Stimulus, it returns NaN, which becomes null when serialized.

With this change, the attribute is now only set if chunk_size is set on the panel, and the request body only includes chunk_size in the arguments if the value exists on the controller